# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "4Foods API",
    "description": "API documentation for 4Foods Android app",
    "version": "1.0.0"
  },
  "host": "api.4foods.app",
  "basePath": "/",
  "schemes": ["https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your bearer token in the format **Bearer <token>**"
    }
  },
  "paths": {
...
    "/user": {
      "get": {
        "tags": ["User"],
        "summary": "List all users (Admin only)",
        "description": "Admin fetches all users.",
        "security": [{ "Bearer": [] }],
        "responses": {
          "200": {
            "description": "List of users",
            "examples": {
              "application/json": [
                {
                  "_id": "6754b6f82e9c2a4f77f1b5d1",
                  "name": "Nguyễn Văn A",
                  "email": "a@example.com",
                  "role": "user",
                  "isSeller": false,
                  "coin": 0
                }
              ]
            }
          },
          "403": {
            "description": "Forbidden",
            "examples": { "application/json": { "error": "Admin access only" } }
          },
          "500": {
            "description": "Server error",
            "examples": {
              "application/json": { "error": "Internal server error" }
            }
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "tags": ["User"],
        "summary": "Get user by ID (Admin only)",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "example": "6754b6f82e9c2a4f77f1b5d1"
          }
        ],
        "responses": {
          "200": {
            "description": "User detail",
            "examples": {
              "application/json": {
                "_id": "6754b6f82e9c2a4f77f1b5d1",
                "name": "Nguyễn Văn A",
                "email": "a@example.com",
                "role": "user",
                "isSeller": false,
                "phone": "0987654321",
                "addresses": [],
                "coin": 0,
                "createdAt": "2025-11-20T08:10:00.000Z"
              }
            }
          },
          "404": {
            "description": "Not found",
            "examples": { "application/json": { "error": "User not found" } }
          },
          "403": {
            "description": "Forbidden",
            "examples": { "application/json": { "error": "Admin access only" } }
          }
        }
      },
      "put": {
        "tags": ["User"],
        "summary": "Update user by ID (Admin only)",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "example": "6754b6f82e9c2a4f77f1b5d1"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "phone": { "type": "string" },
                "role": { "type": "string", "enum": ["user", "admin"] },
                "isSeller": { "type": "boolean" },
                "coin": { "type": "number" }
              }
            },
            "example": {
              "name": "Nguyễn Văn A (Cập nhật)",
              "phone": "0912345678",
              "role": "user",
              "isSeller": true,
              "coin": 120
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "examples": {
              "application/json": {
                "message": "User updated",
                "user": {
                  "_id": "6754b6f82e9c2a4f77f1b5d1",
                  "name": "Nguyễn Văn A (Cập nhật)",
                  "phone": "0912345678",
                  "role": "user",
                  "isSeller": true,
                  "coin": 120
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "examples": { "application/json": { "error": "User not found" } }
          },
          "403": {
            "description": "Forbidden",
            "examples": { "application/json": { "error": "Admin access only" } }
          },
          "500": {
            "description": "Server error",
            "examples": {
              "application/json": { "error": "Internal server error" }
            }
          }
        }
      },
      "delete": {
        "tags": ["User"],
        "summary": "Delete user by ID (Admin only)",
        "description": "Requires a reason in body. Sends email before deletion.",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "example": "6754b6f82e9c2a4f77f1b5d1"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["reason"],
              "properties": { "reason": { "type": "string" } }
            },
            "example": { "reason": "Vi phạm điều khoản sử dụng" }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "examples": {
              "application/json": { "message": "User deleted and email sent" }
            }
          },
          "400": {
            "description": "Missing reason",
            "examples": {
              "application/json": {
                "error": "Reason is required to delete account"
              }
            }
          },
          "404": {
            "description": "Not found",
            "examples": { "application/json": { "error": "User not found" } }
          },
          "403": {
            "description": "Forbidden",
            "examples": { "application/json": { "error": "Admin access only" } }
          },
          "500": {
            "description": "Server error",
            "examples": {
              "application/json": { "error": "Internal server error" }
            }
          }
        }
      }
    },
    "/user/me": {
      "get": {
        "tags": ["User"],
        "summary": "Get current user",
        "security": [{ "Bearer": [] }],
        "responses": {
          "200": {
            "description": "User",
            "examples": {
              "application/json": {
                "id": "6754b6f82e9c2a4f77f1b5d1",
                "name": "Nguyễn Văn A",
                "email": "a@example.com",
                "role": "user",
                "isSeller": false,
                "addresses": []
              }
            }
          }
        }
      },
      "put": {
        "tags": ["User"],
        "summary": "Update current user",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": { "type": "object" },
            "example": {
              "name": "Nguyễn Văn A",
              "phone": "0987654321",
              "dob": "2001-03-15",
              "permanentAddress": "Hà Nội"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "examples": {
              "application/json": {
                "name": "Nguyễn Văn A",
                "phone": "0987654321"
              }
            }
          }
        }
      }
    },
    "/user/me/change-password": {
      "put": {
        "tags": ["User"],
        "summary": "Change password",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "required": ["oldPassword", "newPassword"]
            },
            "example": { "oldPassword": "123456", "newPassword": "Abc@123456" }
          }
        ],
        "responses": {
          "200": {
            "description": "Changed",
            "examples": {
              "application/json": { "message": "Password changed" }
            }
          },
          "400": {
            "description": "Error",
            "examples": {
              "application/json": { "error": "Old password incorrect" }
            }
          }
        }
      }
    },
    "/user/me/addresses": {
      "post": {
        "tags": ["User"],
        "summary": "Add address",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "required": [
                "name",
                "phone",
                "province",
                "district",
                "ward",
                "detail"
              ]
            },
            "example": {
              "label": "home",
              "name": "Nguyễn Văn A",
              "phone": "0912345678",
              "province": "Hà Nội",
              "district": "Cầu Giấy",
              "ward": "Dịch Vọng",
              "detail": "Số 10, Ngõ 20",
              "isPrimary": true
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Added",
            "examples": {
              "application/json": [
                { "name": "Nguyễn Văn A", "isPrimary": true }
              ]
            }
          },
          "400": {
            "description": "Missing",
            "examples": {
              "application/json": { "error": "Missing required address fields" }
            }
          }
        }
      }
    },
    "/user/me/addresses/{idx}": {
      "put": {
        "tags": ["User"],
        "summary": "Update address by index",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "name": "idx",
            "in": "path",
            "required": true,
            "type": "integer",
            "example": 0
          },
          {
            "in": "body",
            "name": "body",
            "schema": { "type": "object" },
            "example": { "note": "Giao sáng", "isPrimary": true }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "examples": {
              "application/json": [
                { "name": "Nguyễn Văn A", "isPrimary": true }
              ]
            }
          },
          "404": {
            "description": "Not found",
            "examples": { "application/json": { "error": "Address not found" } }
          }
        }
      },
      "delete": {
        "tags": ["User"],
        "summary": "Delete address by index",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "name": "idx",
            "in": "path",
            "required": true,
            "type": "integer",
            "example": 0
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "examples": { "application/json": [] }
          },
          "404": {
            "description": "Not found",
            "examples": { "application/json": { "error": "Address not found" } }
          }
        }
      }
    },
    "/user/me/delete-request": {
      "post": {
        "tags": ["User"],
        "summary": "Request account deletion",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": { "type": "object", "required": ["reason"] },
            "example": { "reason": "Không muốn sử dụng nữa" }
          }
        ],
        "responses": {
          "200": {
            "description": "Requested",
            "examples": {
              "application/json": {
                "message": "Yêu cầu xóa tài khoản đã được gửi đến admin"
              }
            }
          }
        }
      }
    },
    "/user/follow/{shopId}": {
      "post": {
        "tags": ["User"],
        "summary": "Follow shop",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "name": "shopId",
            "in": "path",
            "required": true,
            "type": "string",
            "example": "6733f0d024f23c6f9c28a77d"
          }
        ],
        "responses": {
          "200": {
            "description": "Followed",
            "examples": {
              "application/json": { "message": "Shop followed successfully" }
            }
          },
          "404": {
            "description": "Not found",
            "examples": {
              "application/json": { "error": "Shop not found or unavailable" }
            }
          }
        }
      }
    },
    "/user/unfollow/{shopId}": {
      "delete": {
        "tags": ["User"],
        "summary": "Unfollow shop",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "name": "shopId",
            "in": "path",
            "required": true,
            "type": "string",
            "example": "6733f0d024f23c6f9c28a77d"
          }
        ],
        "responses": {
          "200": {
            "description": "Unfollowed",
            "examples": {
              "application/json": { "message": "Unfollowed shop successfully" }
            }
          },
          "404": {
            "description": "Not found",
            "examples": { "application/json": { "error": "Shop not found" } }
          }
        }
      }
    },
    "/user/me/followed-shops": {
      "get": {
        "tags": ["User"],
        "summary": "List followed shops",
        "security": [{ "Bearer": [] }],
        "responses": {
          "200": {
            "description": "List",
            "examples": {
              "application/json": [
                {
                  "_id": "6733f0d024f23c6f9c28a77d",
                  "name": "Tiệm Trà Sữa Posu",
                  "isOpen": true
                }
              ]
            }
          }
        }
      }
    },
...
  }
}